fix: Implemented filtering to exclude non-existent secured variables from the delete request

+++ src/main/java/org/qubership/integration/platform/variables/management/service/SecuredVariableService.java
@@ -157,6 +157,13 @@ public class SecuredVariableService {
         lock.lock();
         try {
             existedVariables = secretService.getSecretData(secretName, true).keySet();
+            variablesNames = variablesNames.stream()
+                    .filter(existedVariables::contains)
+                    .collect(Collectors.toSet());
+            if (variablesNames.isEmpty()) {
+                return;
+            }
+
             secretService.removeEntries(secretName, variablesNames);
         } finally {
             lock.unlock();
@@ -178,9 +185,15 @@ public class SecuredVariableService {
         try {
             variablesBySecret = secretService.getAllSecretsData();
             variablesPerSecret.forEach((secretName, variablesToRemove) -> {
-                boolean secretExists = variablesBySecret.containsKey(secretName);
-                if (!secretExists) {
+                Map<String, String> secretVariables = variablesBySecret.get(secretName);
+                if (secretVariables == null) {
                     secretUpdateExceptions.put(secretName, SecretNotFoundException.forSecret(secretName));
+                    return;
+                }
+                variablesToRemove = variablesToRemove.stream()
+                        .filter(secretVariables::containsKey)
+                        .collect(Collectors.toSet());
+                if (variablesToRemove.isEmpty()) {
                     return;
                 }
